Debugging binary tree

+++ 14-binary_tree_balance.c
@@ -5,7 +5,7 @@
  * @tree: pointer to the root node
  * Return: the measure else 0
  */
-size_t binary_tree_balance(const binary_tree_t *tree)
+int binary_tree_balance(const binary_tree_t *tree)
 {
 	size_t x;
 	size_t y;
